+++ src/pde/kse.rs
@@ -1,43 +1,9 @@
-use fftw::array::*;
-use fftw::plan::*;
-use fftw::types::*;
+use fftw::types::c64;
 use ndarray::*;
 use std::f64::consts::PI;
 
+use super::Pair;
 use traits::*;
-
-/// Pair of one-dimensional Real/Complex aligned arrays
-pub struct Pair {
-    pub r: AlignedVec<f64>,
-    pub c: AlignedVec<c64>,
-    r2c: R2CPlan64,
-    c2r: C2RPlan64,
-}
-
-impl Pair {
-    pub fn new(n: usize) -> Self {
-        let nf = n / 2 + 1;
-        let mut r = AlignedVec::new(n);
-        let mut c = AlignedVec::new(nf);
-        let r2c = R2CPlan::new(&[n], &mut r, &mut c, Flag::Measure).unwrap();
-        let c2r = C2RPlan::new(&[n], &mut c, &mut r, Flag::Measure).unwrap();
-        Pair { r, c, r2c, c2r }
-    }
-
-    pub fn r2c(&mut self) {
-        self.r2c.r2c(&mut self.r, &mut self.c).unwrap();
-    }
-
-    pub fn c2r(&mut self) {
-        self.c2r.c2r(&mut self.c, &mut self.r).unwrap();
-    }
-}
-
-impl Clone for Pair {
-    fn clone(&self) -> Self {
-        Pair::new(self.r.len())
-    }
-}
 
 pub struct KSE {
     n: usize,
@@ -95,10 +61,7 @@ impl SemiImplicit for KSE {
             *u = -*u * *ux;
         }
         self.u.r2c();
-        let normalize = 1.0 / self.n as f64;
-        for (u_, u) in uf.iter_mut().zip(self.u.c.iter()) {
-            *u_ = *u * normalize;
-        }
+        uf.as_slice_mut().unwrap().copy_from_slice(&self.u.c);
         uf
     }
